+++ ms/uplink.c
@@ -99,7 +99,7 @@ void OPENSSL_Uplink(volatile void **table, int index)
     table[index] = func;
 }
 
-#if (defined(_MSC_VER) || defined(__BORLANDC__)) && defined(_M_IX86)
+#if defined(_MSC_VER && defined(_M_IX86)
 # if defined(_MSC_VER)
 #  define LAZY(i)         \
 __declspec(naked) static void lazy##i (void) {  \
@@ -108,7 +108,7 @@ __declspec(naked) static void lazy##i (void) {  \
         _asm    call OPENSSL_Uplink             \
         _asm    add  esp,8                      \
         _asm    jmp  OPENSSL_UplinkTable+4*i    }
-# elif defined(__BORLANDC__) && defined(__clang__)
+# elif defined(__clang__)
 void *OPENSSL_UplinkTable[26]; /* C++Builder requires declaration before use */
 #  define LAZY(i)         \
 __declspec(naked) static void lazy##i (void) { \